refactor: moving define functions into dispatchers package
